--- a/crates/containerd-shim-wasm/src/sys/unix/container/mod.rs
+++ b/crates/containerd-shim-wasm/src/sys/unix/container/mod.rs
@@ -1,2 +1,5 @@
+#[allow(clippy::module_inception)]
+mod container;
+
 mod executor;
 pub mod instance;