@Mossaka 2025-01-30
Great performance increase! I just ran a stress test on 1000 containers and with concurrency=32

**Before**
999 tasks succeeded, 0 tasks failed, 1 tasks didn't finish
elapsed time: 5s 898ms 226us 35ns
 
**After**
1000 tasks succeeded, 0 tasks failed, 0 tasks didn't finish
elapsed time: 2s 48ms 936us 565ns